Add files via upload
This program is a dice-rolling game where two players compete for the highest face value of two dice, one die has 6 sides, and the other die has 10 sides.
The program contains three classes, Die, Player, and HighTwoGame. The HighTwoGame class determines the winner of each game and returns the results.